fix(builder-audit): respect architecture boundaries flagged by CI
- MVT-01: revert the tenant-agnostic tile layer-name change; the dormant-
  tenancy isolation guard (DP-02/T1G) requires the layer name to stay
  schema-qualified. In single_tenant (the only OSS mode) it is already
  'data.{table}'; the multi_tenant client divergence is a deferred cloud-
  overlay concern (client-side source-layer derivation), now documented.
- P1-13: lazy-import filter_grammar inside chat_validation (processing must
  not module-level-import catalog - Phase 225 PROCESS-02 guard).
- P0-01: move revoke_embed_tokens_for_dropped_datasets from maps/service_public
  to embed_tokens/service (next to revoke_embed_tokens_by_map). Fixes the
  facade-boundary violation (router imported a private maps service module)
  and the service_public size-budget overflow at once.
